OcrController add
