+++ project/OnTeacherProject/src/main/webapp/upload/result.txt
@@ -1,1 +1,1 @@
-<보 기> 「최고운전」은 비범한 인물로서의 최치원을 형상화했다. 주인공 은 문제 해결의 국면에서 치밀함, 기지, 당당함을 보인다. 또한 초월적 존재의 도움을 받으면서도 이에 전적으로 의존하지 않고 자신이 지닌 신이한 능력을 발휘하여 개인의 문제와 국가의 과제를 직접 해결한다. 이는 당대 독자들이 원했던 새로운 영웅상을 최치원에 투영하여 작품 속에서 구현한 것이다. 
+<보 기> 우리나라에서는 자기가 사는 지역에서 많이 나는 산물을 다른 데서 산출되는 필요한 물건과 교환하여 풍족하게 살려는 백성이 많으나 힘이 미치지 못한다. ... 중국 사람은 가난하면 장사를 한다. 그렇더라도 정말 사람만 현명하면 원래 가진 풍류와 명망은 그대로다. 그래서 유생이 거리낌 없이 서점을 출입하고, 재상조차도 직접 융복사 앞 시장에 가서 골동품을 산다. *** 우리나라는 해마다 은 수만 냥을 연경에 실어 보내 약재와 비단을 사 오는 반면, 우리나라 물건을 팔아 저들의 은으로 바꿔 오는 일은 없다. 은이란 천년이 지나도 없어지지 않는 물건이지만, 약은 사람에게 먹여 반나절이면 사라져 버리고 비단은 시신을 감싸서 묻으면 반년 만에 썩어 없어진다.
